Modify / Replace RankingTuple with RankingData
